% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboost.R
\name{bayesboost}
\alias{bayesboost}
\title{BayesBoost algorithm for linear mixed model}
\usage{
bayesboost(
  y,
  x,
  mstop,
  id,
  id_constant_vars = NULL,
  t,
  lambda = 0.1,
  hampel_k = 7,
  hampel_nsigma = 2,
  pat = 3,
  max_num_ranef = ifelse(!is.null(ranef_var), length(ranef_var),
    ifelse(is.null(pot_ranef_var), 2, length(pot_ranef_var))),
  pot_ranef_var = NULL,
  ranef_var = NULL,
  seed = NULL
)
}
\arguments{
\item{y}{response variable}

\item{x}{explanatory variables, it contains all variables (including cluster-constant ones).}

\item{mstop}{maximal stopping iteration.}

\item{id}{vector of individuals or clusters.}

\item{id_constant_vars}{cluster-constant variables.}

\item{t}{number of MCMC samplings.}

\item{lambda}{step length or learning rate, default is 0.1.}

\item{hampel_k}{window-size of Hampel-filter, default is 7, see ?hampel.}

\item{hampel_nsigma}{number of standard devisions, default is 2, see ?hampel.}

\item{pat}{"patience" parameter in early stopping, default is 3, see ?earlyStopping.}

\item{max_num_ranef}{Maximal number of random effect that can be included in the final model,
if pot_ranef_var is not NULL, then its default values is the length of pot_ranef_var, else is 2.
if max_num_ranef = 0, then execute random intercept model.}

\item{pot_ranef_var}{A sequence of indexes indicates the potential variables that can be modeled as random effects.}

\item{ranef_var}{predefined random effects structures, subset of x.}

\item{seed}{random.seed for reproduction.}
}
\description{
It is used to perform Bayesian boosting estimation for linear mixed models.
  Stopping iteration is determined according to the cAIC-based stopping criteria with the
  help of Hampel-filter. It provides flexible random effects structures, including the
  totally automatic selection of random effects, automatic selection from a potential
  random effects set and specific random effect structure.
}
