% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAIC.R
\name{cAIC}
\alias{cAIC}
\title{cAIC}
\usage{
cAIC(y, y_hat, X, sigma2, G, U, q, m)
}
\arguments{
\item{y}{response variable}

\item{y_hat}{fitted values of the response variable}

\item{X}{design matrix of fixed effects}

\item{sigma2}{variance of epsilon}

\item{G}{block-diagonal matrix of the var-cov matrix of random effects}

\item{U}{design matrix of random effects}

\item{q}{number of random effects variables without intercept}

\item{m}{number of individuals or clusters}
}
\description{
cAIC
}
