% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboost.R
\name{cls}
\alias{cls}
\title{Componentwise least square}
\usage{
cls(y, x, hat_mat_list)
}
\arguments{
\item{y}{y}

\item{x}{x}

\item{hat_mat_list}{list of hat matrix}
}
\description{
Componentwise least square
}
\keyword{internal}
