% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hampel.R
\name{hampel}
\alias{hampel}
\title{Hampel Filter}
\usage{
hampel(x, k = 7, nsigma = 2)
}
\arguments{
\item{x}{a vector of data points.}

\item{k}{the window size, default is 7.}

\item{nsigma}{a number of standard deviations, default is 2.}
}
\description{
hampel(x, k = 7, nsigma = 2) applies a Hampel filter to the input vector, x, to detect and remove outliers.
For each sample of x, the function computes the median of a window composed of the sample and its six
surrounding samples, three per side. It also estimates the standard deviation of each sample about its
window median using the median absolute deviation. If a sample differs from the median by more than
two standard deviations, it is replaced with the median. If x is a matrix, then hampel treats each
column of x as an independent channel.
}
