% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_beta}
\alias{sample_beta}
\title{Sample beta from multivariate normal distribution}
\usage{
sample_beta(y, X, sigma2, Z, gamma)
}
\arguments{
\item{y}{y}

\item{X}{X}

\item{sigma2}{sigma2}

\item{Z}{design Z matrix}

\item{gamma}{gamma}
}
\description{
Sample beta from multivariate normal distribution
}
\keyword{internal}
