% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_gamma}
\alias{sample_gamma}
\title{Sample gamma from multivariate normal distribution}
\usage{
sample_gamma(Z, r, G, y, X, beta, crossprod_Z)
}
\arguments{
\item{Z}{Z}

\item{r}{r}

\item{G}{G}

\item{y}{y}

\item{X}{X}

\item{beta}{beta}

\item{crossprod_Z}{crossprod_Z}
}
\description{
Sample gamma from multivariate normal distribution
}
\keyword{internal}
