% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_Q}
\alias{sample_Q}
\title{Sample Q from inverse Wishart distribution}
\usage{
sample_Q(v0, m, gamma, Lambda0)
}
\arguments{
\item{v0}{v0}

\item{m}{m}

\item{gamma}{gamma}

\item{Lambda0}{Lambda0}
}
\description{
Sample Q from inverse Wishart distribution
}
\keyword{internal}
