% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_sigma2}
\alias{sample_sigma2}
\title{Sample sigma2 from inverse Gamma distribution}
\usage{
sample_sigma2(a, b, y, X, beta, Z, gamma)
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{y}{y}

\item{X}{X}

\item{beta}{beta}

\item{Z}{Z}

\item{gamma}{gamma}
}
\description{
Sample sigma2 from inverse Gamma distribution
}
\keyword{internal}
