% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_tau2}
\alias{sample_tau2}
\title{Sample tau2 from inverse Gamma distribution}
\usage{
sample_tau2(a_q, b_q, m, gamma)
}
\arguments{
\item{a_q}{a_q}

\item{b_q}{b_q}

\item{m}{m}

\item{gamma}{gamma}
}
\description{
Sample tau2 from inverse Gamma distribution
}
\keyword{internal}
