% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{X_i_fun}
\alias{X_i_fun}
\title{Get cluster design matrix X_i of the fixed effects}
\usage{
X_i_fun(X, n_i)
}
\arguments{
\item{X}{data matrix without intercept}

\item{n_i}{number of observations in each cluster}
}
\description{
Get cluster design matrix X_i of the fixed effects
}
\keyword{internal}
