% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{y_i_fun}
\alias{y_i_fun}
\title{Get clusterwise response y_i}
\usage{
y_i_fun(y, n_i)
}
\arguments{
\item{y}{response vector}

\item{n_i}{number of observations in each cluster}
}
\description{
Get clusterwise response y_i
}
\keyword{internal}
