% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Z_i_fun}
\alias{Z_i_fun}
\title{Get cluster design matrix Z_i of the random effects}
\usage{
Z_i_fun(random_vars, n_i)
}
\arguments{
\item{random_vars}{random variable matrix without intercept}

\item{n_i}{number of observations in each cluster}
}
\description{
Get cluster design matrix Z_i of the random effects
}
\keyword{internal}
